refactor: consolidate Chinese content checking into a single script
- Updated the GitHub Actions workflow to point to the new unified Chinese content checker script.
- Enhanced the checker script to include more robust detection of Chinese content, excluding common false positives.
- Removed the old frontend-specific checker script to streamline the process.
- Added command-line arguments for more flexible usage of the checker.
